+++ src/cli/actions/mod.rs
@@ -1,5 +1,12 @@
-use std::collections::BTreeMap;
-use std::path::PathBuf;
+pub mod acl;
+pub mod bucket;
+pub mod object_delete;
+pub mod object_get;
+pub mod object_list;
+pub mod object_put;
+pub mod object_share;
+
+use std::{collections::BTreeMap, path::PathBuf};
 
 #[derive(Debug)]
 pub enum Action {
@@ -7,14 +14,25 @@ pub enum Action {
         key: String,
         acl: Option<String>,
     },
+    CreateBucket {
+        acl: String,
+    },
+    DeleteObject {
+        key: String,
+        upload_id: String,
+        bucket: bool,
+    },
     ListObjects {
         bucket: Option<String>,
         list_multipart_uploads: bool,
         prefix: Option<String>,
         start_after: Option<String>,
     },
-    MakeBucket {
-        acl: String,
+    GetObject {
+        key: String,
+        get_head: bool,
+        dest: Option<String>,
+        quiet: bool,
     },
     PutObject {
         acl: Option<String>,
@@ -25,16 +43,7 @@ pub enum Action {
         pipe: bool,
         s3m_dir: PathBuf,
         quiet: bool,
-    },
-    DeleteObject {
-        key: String,
-        upload_id: String,
-    },
-    GetObject {
-        key: String,
-        get_head: bool,
-        dest: Option<String>,
-        quiet: bool,
+        tmp_dir: PathBuf,
     },
     ShareObject {
         key: String,