+++ src/cli/mod.rs
@@ -1,24 +1,12 @@
 mod config;
 pub use self::config::{Config, Host};
 
-mod upload;
-pub use self::upload::upload;
-
-mod multipart_upload;
-pub use self::multipart_upload::multipart_upload;
-
-mod db;
-pub use self::db::Db;
-
-mod part;
-pub use self::part::Part;
-
-mod start;
-pub use self::start::start;
-
 pub mod actions;
 pub mod progressbar;
 
 mod commands;
 mod dispatch;
 mod matches;
+
+mod start;
+pub use self::start::start;