+++ src/s3/actions/createmultipartupload.rs
@@ -62,7 +62,7 @@ impl<'a> CreateMultipartUpload<'a> {
     pub async fn request(&self, s3: &S3) -> Result<InitiateMultipartUploadResult> {
         let (url, headers) = &self.sign(s3, tools::sha256_digest("").as_ref(), None, None)?;
         let response =
-            request::request(url.clone(), self.http_method(), headers, None, None).await?;
+            request::request(url.clone(), self.http_method()?, headers, None, None).await?;
 
         if response.status().is_success() {
             let upload_req: InitiateMultipartUploadResult = from_str(&response.text().await?)?;
@@ -75,8 +75,8 @@ impl<'a> CreateMultipartUpload<'a> {
 
 // <https://docs.aws.amazon.com/AmazonS3/latest/API/API_CreateMultipartUpload.html>
 impl<'a> Action for CreateMultipartUpload<'a> {
-    fn http_method(&self) -> Method {
-        Method::from_bytes(b"POST").unwrap()
+    fn http_method(&self) -> Result<Method> {
+        Ok(Method::from_bytes(b"POST")?)
     }
 
     fn headers(&self) -> Option<BTreeMap<&str, &str>> {
@@ -123,6 +123,6 @@ mod tests {
     #[test]
     fn test_method() {
         let action = CreateMultipartUpload::new("key", None, None);
-        assert_eq!(Method::POST, action.http_method());
+        assert_eq!(Method::POST, action.http_method().unwrap());
     }
 }