+++ src/s3/actions/getobject.rs
@@ -29,7 +29,7 @@ impl<'a> GetObject<'a> {
     pub async fn request(&self, s3: &S3) -> Result<reqwest::Response> {
         let (url, headers) = &self.sign(s3, tools::sha256_digest("").as_ref(), None, None)?;
         let response =
-            request::request(url.clone(), self.http_method(), headers, None, None).await?;
+            request::request(url.clone(), self.http_method()?, headers, None, None).await?;
         if response.status().is_success() {
             Ok(response)
         } else {
@@ -40,8 +40,8 @@ impl<'a> GetObject<'a> {
 
 // https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetObject.html
 impl<'a> Action for GetObject<'a> {
-    fn http_method(&self) -> Method {
-        Method::from_bytes(b"GET").unwrap()
+    fn http_method(&self) -> Result<Method> {
+        Ok(Method::from_bytes(b"GET")?)
     }
 
     fn headers(&self) -> Option<BTreeMap<&str, &str>> {
@@ -86,6 +86,6 @@ mod tests {
     #[test]
     fn test_method() {
         let action = GetObject::new("key");
-        assert_eq!(Method::GET, action.http_method());
+        assert_eq!(Method::GET, action.http_method().unwrap());
     }
 }