+++ src/s3/actions/listbuckets.rs
@@ -23,7 +23,7 @@ impl ListBuckets {
     pub async fn request(&self, s3: &S3) -> Result<ListAllMyBucketsResult> {
         let (url, headers) = &self.sign(s3, tools::sha256_digest("").as_ref(), None, None)?;
         let response =
-            request::request(url.clone(), self.http_method(), headers, None, None).await?;
+            request::request(url.clone(), self.http_method()?, headers, None, None).await?;
         if response.status().is_success() {
             let buckets: ListAllMyBucketsResult = from_str(&response.text().await?)?;
             Ok(buckets)
@@ -35,8 +35,8 @@ impl ListBuckets {
 
 // https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListBuckets.html
 impl Action for ListBuckets {
-    fn http_method(&self) -> Method {
-        Method::from_bytes(b"GET").unwrap()
+    fn http_method(&self) -> Result<Method> {
+        Ok(Method::from_bytes(b"GET")?)
     }
 
     fn headers(&self) -> Option<BTreeMap<&str, &str>> {
@@ -60,6 +60,6 @@ mod tests {
     #[test]
     fn test_method() {
         let action = ListBuckets::new();
-        assert_eq!(Method::GET, action.http_method());
+        assert_eq!(Method::GET, action.http_method().unwrap());
     }
 }