+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 0.6.0 🛺  [Unreleased]
-* Added sub-comand `mb` (Make a bucket)
+## 0.6.0 🛺
+* Added sub-comand `cb` (Create bucket)
+* Added option `-b/--bucket` to remove bucket when using `rm`
+* Added option `-t/--tmp-dir` directory for temporarily storing the stdin buffer.
+* Added option `-s/--show` to list current available hosts
 
 ## 0.5.0 🛵
 * removed ending `/` when listing buckets